AutoDesktopVisionApi/requirements.txt
pillow>=9.0.0
requests>=2.27.1
ultralytics>=8.0.0
numpy>=1.20.0
opencv-python>=4.5.0
scikit-image>=0.19.0
